Quality of life fixes
Add a system that will prevent the temi from moving if a user is close to it. Also changed the close thresh to be further out. Add system to allow the screen to move to a default position, but then to go slack to allow user to move the screen to desired position. Also added a timeout feature that is set to five min that will exit the application if state is other then quizhome.